# .well-known/openapi.yaml
openapi: 3.0.2
info:
  title: OpenAI plugin for a simple grocery list app
  description: A grocery list app for managing your grocery list on ChatGPT
  version: 1.0.0
  servers:
    - url: https://minsa110-sturdy-space-acorn-5gq4prgj9q9c7q45-8000.preview.app.github.dev/
paths:
  /items:
    post:
      summary: Create a new grocery list item
      description: Accepts a string and adds as new grocery list item
      operationId: create_item
      parameters:
        - in: query
          name: item
          schema:
            type: string
          required: true
          description: The description of the item
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    get:
      summary: Get a list of all grocery list items
      operationId: list_items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryList"
  /items/{item_id}:
    get:
      summary: Get an item by ID
      operationId: get_item
      parameters:
        - name: item_id
          in: path
          required: true
          description: ID of the item to retrieve
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Item not found
    delete:
      summary: Delete an item by ID
      operationId: delete_item
      parameters:
        - name: item_id
          in: path
          required: true
          description: ID of the item to delete
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Item deleted
        "404":
          description: Item not found
components:
  schemas:
    Item:
      type: object
      properties:
        item:
          type: string
        item_id:
          type: integer
          format: int32
          readOnly: true
      required:
        - item